Adding Library class fix
