+++ spec/library_spec.rb
@@ -1,10 +1,13 @@
-require "yaml"
+require './lib/libary'
+require './lib/person'
+require 'yaml'
 
-class Library
+describe Library do
 
-  attr_accessor :books
+  let(:person) { instance_double('Person', name: 'Lisa' )}
+  subject { described class.new(name: 'Lisa') }
 
-  def initialize
-    @books = YAML.load_file(".lib/data.yml")
+  it 'is expected to have a record of books on initialize' do
+    expect(subject.books).not_to be_empty
   end
 end
